+++ main.cpp
@@ -5,10 +5,10 @@
 #include <string>
 #include <cstdlib>
 #include <ctime>
-#include "include/DataStructures/DataVector/DataVector.h"
-#include "include/Algorithms/VamanaIndex.h"
+#include "include/DataVector.h"
+#include "include/VamanaIndex.h"
 #include "include/read_data.h"
-#include "include/DataStructures/DataVector/BQDataVectors.h"
+#include "include/BQDataVectors.h"
 
 using ParametersMap = std::map<std::string, std::string>;
 using BaseVectors = std::vector<DataVector<float>>;
