+++ tests/test_data_vectors.cc
@@ -3,7 +3,7 @@
 #include <vector>
 #include "../include/acutest.h"
 #include "../include/read_data.h"
-#include "../include/DataStructures/DataVector/DataVector.h"
+#include "../include/DataVector.h"
 #include <cmath>  // Ensure cmath is included for fabs
 
 /**
